+++ components/hfs-server/src/web/op/store_series_data.js
@@ -8,7 +8,7 @@ const business = require('components/business');
 const ApiConstants = require('../api_constants');
 const TracedOperations = require('./traced_operations');
 
-const setCommonMeta = require('components/api-server/src/methods/helpers/setCommonMeta');
+const setCommonMeta = require('components/api-server/src/methods/helpers/setCommonMeta').setCommonMeta;
 
 import type { TypeRepository } from 'components/business';
 import type Context from '../../context';