+++ components/hfs-server/test/unit/Settings.test.js
@@ -4,10 +4,10 @@
 const should = require('should');
 
 /* global describe, it */
-const { settings } = require('./test-helpers');
+const { loadSettings } = require('./test-helpers');
 
 describe('Settings', function() {
-  it('[KEEZ] should have been loaded for test execution', function() {
-    should.exist(settings);
+  it('[KEEZ] should have been loaded for test execution', async function() {
+    should.exist(await loadSettings());
   });
 });